# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "範例API文件工具.",
        "title": "GoPro API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Zuolar"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/login": {
            "post": {
                "description": "設定cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "登入",
                "parameters": [
                    {
                        "description": "登入資訊",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/main.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登入狀況",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "delete": {
                "description": "清除cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "登出",
                "responses": {
                    "200": {
                        "description": "登出狀況",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/password": {
            "put": {
                "description": "變更自己的密碼",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "變更密碼",
                "parameters": [
                    {
                        "type": "string",
                        "description": "密碼資料",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "變更狀態",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "請先登入",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "description": "取使用者ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "使用者資料",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "使用者ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "使用者資料",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/main.User"
                        }
                    },
                    "401": {
                        "description": "請先登入",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Login": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "qwe123"
                },
                "username": {
                    "type": "string",
                    "example": "zuolar"
                }
            }
        },
        "main.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 99
                },
                "password": {
                    "type": "string",
                    "example": "qwe123"
                },
                "username": {
                    "type": "string",
                    "example": "zuolar"
                }
            }
        }
    }
}